Add zsh syntax highlighting plugin - Refer https://github.com/zsh-users/zsh-syntax-highlighting - Minor install.sh improvements: Check before downloading configuration

+++ install.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 DOTFILES_ROOT=$(pwd -P)
 
-if ! test -f git/.gitconfig.local.symlink; then
+if [ ! -f git/.gitconfig.local.symlink ]; then
     read -p 'What is your git user name?' git_author
     read -p 'What is your personal git email?' git_personal_email
     read -p 'What is your work git email?' git_work_email
@@ -14,8 +14,15 @@ sed -e "s=DOTFILES_ROOT=${DOTFILES_ROOT}=g" zsh/.zshrc.symlink > zsh/.zshrc.loca
 
 # git autocompletion
 # Sourced from https://github.com/git/git/blob/v2.33.0/contrib/completion/git-completion.zsh. Credit to authors.
-curl https://raw.githubusercontent.com/git/git/v2.33.0/contrib/completion/git-completion.bash -o "$DOTFILES_ROOT/zsh/_git-completion-bash"
-curl https://raw.githubusercontent.com/git/git/v2.33.0/contrib/completion/git-completion.zsh -o "$DOTFILES_ROOT/zsh/_git-completion-zsh"
+if [ ! -f "$DOTFILES_ROOT/zsh/_git-completion-bash" ]; then
+    curl https://raw.githubusercontent.com/git/git/v2.33.0/contrib/completion/git-completion.bash -o "$DOTFILES_ROOT/zsh/_git-completion-bash"
+    curl https://raw.githubusercontent.com/git/git/v2.33.0/contrib/completion/git-completion.zsh -o "$DOTFILES_ROOT/zsh/_git-completion-zsh"
+fi
+
+# https://github.com/zsh-users/zsh-syntax-highlighting
+if [ ! -d "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting" ]; then
+    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting"
+fi
 
 ln "$DOTFILES_ROOT/git/.gitconfig.local.symlink" ~/.gitconfig
